Rename cmdlets/models for consistency
- Rename all 12 cmdlet classes and files from *DeploymentStackWhatIf to *DeploymentStackWhatIfResult (match the cmdlet verb-noun name)

- Prefix all 13 WhatIf model classes with PSDeploymentStackWhatIf for consistency (DeploymentExtension -> PSDeploymentStackWhatIfExtension, DeploymentStackChangeBase -> PSDeploymentStackWhatIfChangeBase, etc.)

- Move all 15 WhatIf model files from SdkModels/Deployments/ to new SdkModels/DeploymentStackWhatIf/ directory

- Update namespace from ...SdkModels.Deployments to ...SdkModels.DeploymentStackWhatIf

- Update all references in SDK clients, formatter, and cmdlets
