+++ src/Resources/ResourceManager/Implementation/CmdletBase/DeploymentStackWhatIfCreateCmdlet.cs
@@ -17,6 +17,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Cmdlet
     using System;
     using System.Management.Automation;
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.Deployments;
+    using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.DeploymentStackWhatIf;
 
     /// <summary>
     /// Base class for Deployment Stack What-If cmdlets.
