+++ src/Resources/ResourceManager/Implementation/DeploymentStacksWhatIf/RemoveAzSubscriptionDeploymentStackWhatIfResult.cs
@@ -17,31 +17,25 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Deploy
     using System;
     using System.Management.Automation;
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Components;
-    using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.Deployments;
+    using Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.DeploymentStackWhatIf;
     using Microsoft.Azure.Commands.ResourceManager.Common;
 
     /// <summary>
-    /// Deletes an existing WhatIf result for a Management Group Deployment Stack.
+    /// Deletes an existing WhatIf result for a Subscription Deployment Stack.
     /// </summary>
-    [Cmdlet("Remove", AzureRMConstants.AzureRMPrefix + "ManagementGroupDeploymentStackWhatIfResult",
-        SupportsShouldProcess = true, DefaultParameterSetName = RemoveByNameAndManagementGroupIdParameterSetName), OutputType(typeof(bool))]
-    public class RemoveAzManagementGroupDeploymentStackWhatIf : DeploymentStacksCmdletBase
+    [Cmdlet("Remove", AzureRMConstants.AzureRMPrefix + "SubscriptionDeploymentStackWhatIfResult",
+        SupportsShouldProcess = true, DefaultParameterSetName = RemoveByNameParameterSetName), OutputType(typeof(bool))]
+    public class RemoveAzSubscriptionDeploymentStackWhatIfResult : DeploymentStacksCmdletBase
     {
         internal const string RemoveByResourceIdParameterSetName = "RemoveByResourceId";
-        internal const string RemoveByNameAndManagementGroupIdParameterSetName = "RemoveByNameAndManagementGroupId";
+        internal const string RemoveByNameParameterSetName = "RemoveByName";
         internal const string RemoveByInputObjectParameterSetName = "RemoveByInputObject";
 
-        [Parameter(Position = 1, Mandatory = true, ValueFromPipelineByPropertyName = true,
-            ParameterSetName = RemoveByNameAndManagementGroupIdParameterSetName,
+        [Parameter(Position = 0, Mandatory = true, ValueFromPipelineByPropertyName = true,
+            ParameterSetName = RemoveByNameParameterSetName,
             HelpMessage = "The name of the WhatIf result to delete.")]
         [ValidateNotNullOrEmpty]
         public string Name { get; set; }
-
-        [Parameter(Position = 0, Mandatory = true, ValueFromPipelineByPropertyName = true,
-            ParameterSetName = RemoveByNameAndManagementGroupIdParameterSetName,
-            HelpMessage = "The ID of the management group.")]
-        [ValidateNotNullOrEmpty]
-        public string ManagementGroupId { get; set; }
 
         [Alias("Id")]
         [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true,
@@ -71,36 +65,23 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation.Deploy
                     ResourceId = InputObject.Id;
                 }
 
-                // Resolve Name and ManagementGroupId if ResourceId was provided
-                if (!string.IsNullOrEmpty(ResourceId))
-                {
-                    Name = Name ?? ResourceIdUtility.GetDeploymentName(ResourceId);
-                    if (string.IsNullOrEmpty(ManagementGroupId))
-                    {
-                        // Extract management group id from: /providers/Microsoft.Management/managementGroups/{mgId}/...
-                        var segments = ResourceId.Split('/');
-                        var mgIndex = Array.IndexOf(segments, "managementGroups");
-                        if (mgIndex >= 0 && mgIndex + 1 < segments.Length)
-                        {
-                            ManagementGroupId = segments[mgIndex + 1];
-                        }
-                    }
-                }
+                // Resolve Name if ResourceId was provided
+                Name = Name ?? ResourceIdUtility.GetDeploymentName(ResourceId);
 
-                if (Name == null || ManagementGroupId == null)
+                if (Name == null)
                 {
                     throw new PSArgumentException($"Provided Id '{ResourceId}' is not in correct form. Should be in form " +
-                        "/providers/Microsoft.Management/managementGroups/<mgId>/providers/Microsoft.Resources/deploymentStacksWhatIfResults/<name>");
+                        "/subscriptions/<subid>/providers/Microsoft.Resources/deploymentStacksWhatIfResults/<name>");
                 }
 
                 ConfirmAction(
                     Force.IsPresent,
-                    $"Are you sure you want to delete WhatIf result '{Name}' in ManagementGroup '{ManagementGroupId}'?",
+                    $"Are you sure you want to delete WhatIf result '{Name}' in the current subscription?",
                     "Delete",
                     Name,
                     () =>
                     {
-                        DeploymentStacksWhatIfSdkClient.DeleteManagementGroupDeploymentStackWhatIfResult(ManagementGroupId, Name);
+                        DeploymentStacksWhatIfSdkClient.DeleteSubscriptionDeploymentStackWhatIfResult(Name);
                         if (PassThru.IsPresent)
                         {
                             WriteObject(true);
