+++ src/Resources/ResourceManager/SdkModels/DeploymentStackWhatIf/PSDeploymentStackWhatIfChangeDeltaRecord.cs
@@ -11,19 +11,20 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.Deployments
+namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.DeploymentStackWhatIf
 {
+    using System.Collections.Generic;
     using Newtonsoft.Json;
 
-    public class DeploymentStackChangeBase
+    public class PSDeploymentStackWhatIfChangeDeltaRecord
     {
-        [JsonProperty("changeType")]
-        public string ChangeType { get; set; }
-
         [JsonProperty("before")]
         public object Before { get; set; }
 
         [JsonProperty("after")]
         public object After { get; set; }
+
+        [JsonProperty("delta")]
+        public IList<PSDeploymentStackWhatIfPropertyChange> Delta { get; set; }
     }
 }
