+++ src/Resources/ResourceManager/SdkModels/DeploymentStackWhatIf/PSDeploymentStackWhatIfExtensionInfo.cs
@@ -11,20 +11,23 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.Deployments
+namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.DeploymentStackWhatIf
 {
     using System.Collections.Generic;
     using Newtonsoft.Json;
 
-    public class DeploymentStackChangeDeltaRecord
+    public class PSDeploymentStackWhatIfExtensionInfo
     {
-        [JsonProperty("before")]
-        public object Before { get; set; }
+        [JsonProperty("name")]
+        public string Name { get; set; }
 
-        [JsonProperty("after")]
-        public object After { get; set; }
+        [JsonProperty("version")]
+        public string Version { get; set; }
 
-        [JsonProperty("delta")]
-        public IList<DeploymentStackPropertyChange> Delta { get; set; }
+        [JsonProperty("configId")]
+        public string ConfigId { get; set; }
+
+        [JsonProperty("config")]
+        public IDictionary<string, PSDeploymentStackWhatIfExtensionConfigItem> Config { get; set; }
     }
 }
