+++ src/Resources/ResourceManager/SdkModels/DeploymentStackWhatIf/PSDeploymentStackWhatIfKeyVaultInfo.cs
@@ -11,23 +11,13 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.Deployments
+namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkModels.DeploymentStackWhatIf
 {
-    using System.Collections.Generic;
     using Newtonsoft.Json;
 
-    public class DeploymentExtension
+    public class PSDeploymentStackWhatIfKeyVaultInfo
     {
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("version")]
-        public string Version { get; set; }
-
-        [JsonProperty("configId")]
-        public string ConfigId { get; set; }
-
-        [JsonProperty("config")]
-        public IDictionary<string, DeploymentStackExtensionConfigItem> Config { get; set; }
+        [JsonProperty("id")]
+        public string Id { get; set; }
     }
 }
